PagingResponse

+++ AspTemplate/Responses/PagingResponse.cs
@@ -2,5 +2,5 @@ public class PagingResponse<T> : RangeResponse<T>
 {
     public required int Size;
     public required int Total { get; set; }
-    public int MaxPage { get { return (int)Math.Ceiling(((double)Total / (double)Size)); } }
+    public int MaxPage { get { return (int)Math.Ceiling((double)Total / Size); } }
 }
